+++ src/components/ProfitAndLossDetailReport/ProfitAndLossDetailReport.tsx
@@ -15,7 +15,7 @@ import { Button } from '../ui/Button/Button'
 import { VStack, HStack } from '../ui/Stack/Stack'
 import { Label } from '../ui/Typography/Text'
 import { format } from 'date-fns'
-import { LedgerEntrySourceType } from '../../schemas/ledgerEntry'
+import { LedgerEntrySourceType } from '../../schemas/ledgerEntrySource'
 import { Direction } from '../../types'
 import { BreadcrumbItem, DetailReportBreadcrumb } from '../DetailReportBreadcrumb/DetailReportBreadcrumb'
 import type { PnlDetailLine } from '../../hooks/useProfitAndLoss/useProfitAndLossDetailLines'
