+++ src/features/bankTransactions/[bankTransactionId]/metadata/api/useSetMetadataOnBankTransaction.ts
@@ -4,8 +4,8 @@ import { useLayerContext } from '../../../../../contexts/LayerContext'
 import { patch } from '../../../../../api/layer/authenticated_http'
 import { useCallback } from 'react'
 import { useBankTransactionsInvalidator, useBankTransactionsOptimisticUpdater } from '../../../../../hooks/useBankTransactions/useBankTransactions'
-import { encodeVendor, type VendorSchema } from '../../../../vendors/vendorsSchemas'
-import { encodeCustomer, type CustomerSchema } from '../../../../customers/customersSchemas'
+import { encodeVendor, type VendorSchema } from '../../../../../schemas/vendor'
+import { encodeCustomer, type CustomerSchema } from '../../../../../schemas/customer'
 import { useMinMutatingMutation } from '../../../../../hooks/mutation/useMinMutatingMutation'
 
 const SET_METADATA_ON_BANK_TRANSACTION_TAG_KEY = '#set-metadata-on-bank-transaction'
