+++ src/features/customerVendor/customerVendorSchemas.ts
@@ -1,5 +1,5 @@
-import { CustomerSchema } from '../customers/customersSchemas'
-import { VendorSchema } from '../vendors/vendorsSchemas'
+import { CustomerSchema } from '../../schemas/customer'
+import { VendorSchema } from '../../schemas/vendor'
 import { Schema } from 'effect'
 
 export const CustomerVendorTypeSchema = Schema.Literal('CUSTOMER', 'VENDOR')
