+++ src/features/customers/api/useListCustomers.ts
@@ -4,7 +4,7 @@ import { useLayerContext } from '../../../contexts/LayerContext'
 import { get } from '../../../api/layer/authenticated_http'
 import { toDefinedSearchParameters } from '../../../utils/request/toDefinedSearchParameters'
 import { Schema, pipe } from 'effect'
-import { CustomerSchema } from '../customersSchemas'
+import { CustomerSchema } from '../../../schemas/customer'
 
 const ListCustomersRawResultSchema = Schema.Struct({
   data: Schema.Array(CustomerSchema),
